+++ km_api/know_me/serializers/profile_group_serializers.py
@@ -5,7 +5,7 @@ from rest_framework import serializers
 
 from know_me import models
 
-from .profile_row_serializers import ProfileRowSerializer
+from .profile_topic_serializers import ProfileTopicSerializer
 
 
 class ProfileGroupListSerializer(serializers.HyperlinkedModelSerializer):
@@ -26,22 +26,22 @@ class ProfileGroupDetailSerializer(ProfileGroupListSerializer):
 
     Based off ``ProfileGroupListSerializer``.
     """
-    rows_url = serializers.SerializerMethodField()
-    rows = ProfileRowSerializer(many=True, read_only=True)
+    topics_url = serializers.SerializerMethodField()
+    topics = ProfileTopicSerializer(many=True, read_only=True)
 
     class Meta:
-        fields = ('id', 'url', 'name', 'is_default', 'rows_url', 'rows')
+        fields = ('id', 'url', 'name', 'is_default', 'topics_url', 'topics')
         model = models.ProfileGroup
 
-    def get_rows_url(self, group):
+    def get_topics_url(self, group):
         """
-        Get the URL of the given group's row list.
+        Get the URL of the given group's topic list.
 
         Args:
             group:
                 The group being serialized.
 
         Returns:
-            The URL of the given group's row list.
+            The URL of the given group's topic list.
         """
-        return group.get_row_list_url(self.context['request'])
+        return group.get_topic_list_url(self.context['request'])