+++ km_api/know_me/tests/views/test_profile_item_list_view.py
@@ -6,25 +6,25 @@ from know_me import serializers, views
 profile_item_list_view = views.ProfileItemListView.as_view()
 
 
-def test_anonymous(api_rf, profile_row_factory):
+def test_anonymous(api_rf, profile_topic_factory):
     """
     Anonymous users should not be able to access the view.
     """
-    row = profile_row_factory()
+    topic = profile_topic_factory()
 
-    request = api_rf.get(row.get_item_list_url())
-    response = profile_item_list_view(request, pk=row.pk)
+    request = api_rf.get(topic.get_item_list_url())
+    response = profile_item_list_view(request, pk=topic.pk)
 
     assert response.status_code == status.HTTP_403_FORBIDDEN
 
 
-def test_create(api_rf, media_resource_factory, profile_row_factory):
+def test_create(api_rf, media_resource_factory, profile_topic_factory):
     """
     Sending a POST request to the view with valid data should create a
     new profile item.
     """
-    row = profile_row_factory()
-    group = row.group
+    topic = profile_topic_factory()
+    group = topic.group
     profile = group.profile
     media_resource = media_resource_factory(profile=profile)
 
@@ -36,13 +36,13 @@ def test_create(api_rf, media_resource_factory, profile_row_factory):
         'text': 'Some sample text.',
     }
 
-    request = api_rf.post(row.get_item_list_url(), data)
-    response = profile_item_list_view(request, pk=row.pk)
+    request = api_rf.post(topic.get_item_list_url(), data)
+    response = profile_item_list_view(request, pk=topic.pk)
 
     assert response.status_code == status.HTTP_201_CREATED
 
     serializer = serializers.ProfileItemSerializer(
-        row.items.get(),
+        topic.items.get(),
         context={
             'profile': profile,
             'request': request,
@@ -51,27 +51,27 @@ def test_create(api_rf, media_resource_factory, profile_row_factory):
     assert response.data == serializer.data
 
 
-def test_get_items(api_rf, profile_item_factory, profile_row_factory):
+def test_get_items(api_rf, profile_item_factory, profile_topic_factory):
     """
     This view should return a serialized list of profile items belonging
-    to the given row.
+    to the given topic.
     """
-    row = profile_row_factory()
-    profile_item_factory(row=row)
-    profile_item_factory(row=row)
+    topic = profile_topic_factory()
+    profile_item_factory(topic=topic)
+    profile_item_factory(topic=topic)
 
-    group = row.group
+    group = topic.group
     profile = group.profile
 
     api_rf.user = profile.user
 
-    request = api_rf.get(row.get_item_list_url())
-    response = profile_item_list_view(request, pk=row.pk)
+    request = api_rf.get(topic.get_item_list_url())
+    response = profile_item_list_view(request, pk=topic.pk)
 
     assert response.status_code == status.HTTP_200_OK
 
     serializer = serializers.ProfileItemSerializer(
-        row.items.all(),
+        topic.items.all(),
         context={'request': request},
         many=True)